Add files via upload
An updated version of the Assignment 3 file with the generated pairplot. I cannot be bothered to wait a full minute for it to generate when refreshing the kernel at this time. Moving onto step 2 from now on, will need to update the markdowns for our observations of the heatmap and pairplot.